ai-services/contact-sage/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
python-dotenv
httpx
tenacity
python-jose[cryptography]
passlib[bcrypt]
thefuzz
PyJWT
cryptography
sqlalchemy
opentelemetry-instrumentation-sqlalchemy